requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
flask-cors==4.0.0
pyodbc==5.0.1
SQLAlchemy==2.0.25
Werkzeug==2.3.7
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.2
pytz==2024.1
pytest>=7.2,<8.0
pytest-flask>=1.2,<2.0
pytest-cov>=4.0,<5.0
gunicorn>=20.1.0,<21.0